Enable telephony audio device for pixel devices.
Pixel devices support a special audio device called TYPE_TELEPHONY which
enables system apps to output audio in the telephony outgoing audio stream.
This is used for some carriers to play a "call recording tone" which the
remote party will hear if the user starts recording audio with a call
recording app.

Test: Manual testing
Bug: 64138141
Change-Id: I6e1981a4baaea34c0cea54f1af7aa2945c6e09a6